retriable fetches for orders requests
