+++ src/services/actions/thunk/order.js
@@ -2,6 +2,7 @@ import { API_URL } from '../../../utils/config'
 import { GET_ORDER_REQUEST, GET_ORDER_SUCCESS, GET_ORDER_FAILED } from '../order'
 import { GET_ORDERS_REQUEST, GET_ORDERS_SUCCESS, GET_ORDERS_FAILED } from '../orders'
 import { getCookie } from '../../../utils/helpers'
+import { retriableFetch } from '../../../utils/api'
 
 export const order = items => {
   return async function (dispatch) {
@@ -11,7 +12,7 @@ export const order = items => {
     })
     try {
       const ids = items.map(el => el._id)
-      const res = await fetch(API_URL + 'orders', {
+      const res = await retriableFetch(API_URL + 'orders', {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
@@ -19,18 +20,10 @@ export const order = items => {
         },
         body: JSON.stringify({ ingredients: ids })
       })
-      if (res && res.ok) {
-        const resData = await res.json()
-        dispatch({
-          type: GET_ORDER_SUCCESS,
-          orderNumber: resData.order.number
-        })
-      } else {
-        dispatch({
-          type: GET_ORDER_FAILED,
-          payload: { message: 'Ошибка при создании заказа. Пожалуйста, повторите позже.' }
-        })
-      }
+      dispatch({
+        type: GET_ORDER_SUCCESS,
+        orderNumber: res.order.number
+      })
     } catch (e) {
       dispatch({
         type: GET_ORDER_FAILED,
@@ -48,25 +41,17 @@ export const getOrders = () => {
       type: GET_ORDERS_REQUEST
     })
     try {
-      const res = await fetch(API_URL + 'orders', {
+      const res = await retriableFetch(API_URL + 'orders', {
         method: 'GET',
         headers: {
           'Content-Type': 'application/json',
-          Authorization: 'Bearer ' + token
+          authorization: 'Bearer ' + token
         }
       })
-      if (res && res.ok) {
-        const resData = await res.json()
-        dispatch({
-          type: GET_ORDERS_SUCCESS,
-          data: resData.orders
-        })
-      } else {
-        dispatch({
-          type: GET_ORDERS_FAILED,
-          payload: { message: 'Ошибка при получении списка заказов. Пожалуйста, повторите позже.' }
-        })
-      }
+      dispatch({
+        type: GET_ORDERS_SUCCESS,
+        data: res.orders
+      })
     } catch (e) {
       dispatch({
         type: GET_ORDERS_FAILED,
